requirements.txt
paho-mqtt
ConfigArgParse
yoyo-migrations
systemd-python; sys_platform == "linux"
aiohttp
aiohttp_jinja2
jinja2
uvloop; sys_platform == "linux"
cchardet
aiohttp_debugtoolbar

pylint
autopep8
flake8
setuptools
wheel
twine
